Updating locally

+++ lib/Rex/Commands/MD5.pm
@@ -24,8 +24,6 @@ This is just a helper function and will not be reported.
 
 =cut
 
-
-
 package Rex::Commands::MD5;
 
 use strict;
@@ -37,7 +35,6 @@ use Rex::Interface::Exec;
 use Rex::Interface::File;
 use Rex::Interface::Fs;
 use Rex::Helper::Path;
-
 
 require Rex::Exporter;
 use base qw(Rex::Exporter);
@@ -59,42 +56,49 @@ sub md5 {
   my ($file) = @_;
 
   my $fs = Rex::Interface::Fs->create;
-  if($fs->is_file($file)) {
+  if ( $fs->is_file($file) ) {
 
     Rex::Logger::debug("Calculating Checksum (md5) of $file");
-
-    my $script = q|
-    use Digest::MD5;
-    print Digest::MD5::md5_hex(<>) . "\n";
-    |;
-
-    my $rnd_file = get_tmp_file;
-
-    my $fh = Rex::Interface::File->create;
-    $fh->open(">", $rnd_file);
-    $fh->write($script);
-    $fh->close;
 
     my $exec = Rex::Interface::Exec->create;
     my $md5;
 
-    if(Rex::is_local() && $^O =~ m/^MSWin/) {
-      $md5 = $exec->exec("perl $rnd_file \"$file\"");
+    my $os = $exec->exec("uname -s");
+    if ( $os =~ /bsd/i ) {
+      $md5 = $exec->exec("/sbin/md5 -q '$file'");
     }
     else {
-      $md5 = $exec->exec("perl $rnd_file '$file'");
+      ($md5) = split( /\s/, $exec->exec("md5sum '$file'") );
     }
 
-    unless($? == 0) {
-      ($md5) = split(/\s/, $exec->exec("md5sum '$file'"));
-    }
+    if ( $? != 0 ) {
 
-    unless($? == 0) {
-      Rex::Logger::info("Unable to get md5 sum of $file");
-      die("Unable to get md5 sum of $file");
-    }
+      my $script = q|
+      use Digest::MD5;
+      print Digest::MD5::md5_hex(<>) . "\n";
+      |;
 
-    Rex::Interface::Fs->create->unlink($rnd_file);
+      my $rnd_file = get_tmp_file;
+
+      my $fh = Rex::Interface::File->create;
+      $fh->open( ">", $rnd_file );
+      $fh->write($script);
+      $fh->close;
+
+      if ( Rex::is_local() && $^O =~ m/^MSWin/ ) {
+        $md5 = $exec->exec("perl $rnd_file \"$file\"");
+      }
+      else {
+        $md5 = $exec->exec("perl $rnd_file '$file'");
+      }
+
+      unless ( $? == 0 ) {
+        Rex::Logger::info("Unable to get md5 sum of $file");
+        die("Unable to get md5 sum of $file");
+      }
+
+      Rex::Interface::Fs->create->unlink($rnd_file);
+    }
 
     Rex::Logger::debug("MD5SUM ($file): $md5");
     $md5 =~ s/[\r\n]//gms;
